Create BooksPage component with styles
