+++ src/book/pages/BooksPage/BooksPage.css
@@ -1,0 +1,16 @@
+.page-header__title,
+.page-header__counter {
+  font-family: var(--title-font);
+  color: var(--dark-color);
+  font-weight: 400;
+}
+
+.page-header__title {
+  font-size: 1.25rem;
+  line-height: 1.25rem;
+  border-bottom: 1px solid var(--dark-color);
+}
+
+.page-header__counter {
+  font-size: 1rem;
+}
