+++ src/components/Layout/Layout.css
@@ -10,6 +10,5 @@
 
 .page-container {
   height: 100%;
-  padding: 0 20px;
-  margin: auto 0;
+  padding: 20px;
 }
