+++ src/pages/NotFoundPage/NotFoundPage.css
@@ -3,7 +3,7 @@
   flex-direction: column;
   align-items: center;
   gap: 10px;
-  margin: auto 0;
+  margin-top: 100px;
 }
 
 .not-found__title {
